+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add AMI reference for flatcar 3227.2.2.
+- Lifecycle hook for launching master instances in HA mode.
 
 ### Changed
 
 - Bump k8scc to 15.0.0.
 - Disable kube-proxy on release v19 and newer.
+- Allow master node to change the autoscaling healthcheck.
 
 ## [13.2.1] - 2022-08-31